# lt-trim: new option --match-section