@mr-martian 2022-09-29
My original design of `cli.h` was with the intent that situations like this could easily have `-s regex1 -s regex2`, and I'm not sure whether this method of supporting both multiple arguments and comma separation is good or bad.